+++ src/main/scala/citeappscalajs/CiteMainController.scala
@@ -1,77 +1,84 @@
 package citeapp
-import com.thoughtworks.binding._
+import com.thoughtworks.binding.{Binding, dom}
+import com.thoughtworks.binding.Binding.{BindingSeq, Var, Vars}
 import scala.scalajs.js
-import org.scalajs.dom.document
+import scala.scalajs.js._
+import js.annotation._
+import scala.concurrent._
+//import ExecutionContext.Implicits.global
+import collection.mutable
+import collection.mutable._
+import scala.scalajs.js.Dynamic.{ global => g }
+import org.scalajs.dom._
+import org.scalajs.dom.ext._
 import org.scalajs.dom.raw._
-import org.scalajs.dom.ext.Ajax
-import scala.concurrent
-.ExecutionContext
-.Implicits
-.global
 import edu.holycross.shot.cite._
 import edu.holycross.shot.scm._
 import edu.holycross.shot.ohco2._
+import edu.holycross.shot.citeobj._
+
+import monix.execution.Scheduler.Implicits.global
+import monix.eval._
+
 import scala.scalajs.js.annotation.JSExport
 
-@JSExport
+@JSExportTopLevel("citeapp.CiteMainController")
 object CiteMainController {
 
 	@JSExport
-	def main(libUrl: String, libDelim: String): Unit = {
+	def main(libUrl: String): Unit = {
+
+//		ImageModel.imgArchivePath = localImagePath
 
 		CiteMainController.updateUserMessage("Loading default library. Please be patient…",1)
-		js.timers.setTimeout(500){ CiteMainController.loadRemoteLibrary(libUrl, libDelim) }
+		val task = Task{ CiteMainController.loadRemoteLibrary(libUrl) }
+		val future = task.runAsync
 
 		dom.render(document.body, CiteMainView.mainDiv)
 	}
 
-	def loadRemoteLibrary(url: String, libDelim: String):Unit = {
+	def loadRemoteLibrary(url: String):Unit = {
 
 		val xhr = new XMLHttpRequest()
 		xhr.open("GET", url )
 		xhr.onload = { (e: Event) =>
 			if (xhr.status == 200) {
 				val contents:String = xhr.responseText
-				CiteMainController.updateRepository(contents, libDelim)
+				CiteMainController.updateRepository(contents)
 			} else {
 				CiteMainController.updateUserMessage(s"Request for remote library failed with code ${xhr.status}",2)
 			}
 		}
 		xhr.send()
 
-		/*
-		Ajax.get(url).onSuccess { case xhr =>
+		/* Ajax.get(url).onSuccess { case xhr =>
 		CiteMainController.updateUserMessage("Got remote library.",0)
 		val contents:String = xhr.responseText
-		CiteMainController.updateRepository(contents, libDelim)
+		CiteMainController.updateRepository(contents, libDelim, fieldDelim)
 	}
 	*/
 }
 
 	def updateUserMessage(msg: String, alert: Int): Unit = {
-		CiteMainModel.userMessageVisibility := "app_visible"
-		CiteMainModel.userMessage := msg
+		CiteMainModel.userMessageVisibility.value = "app_visible"
+		CiteMainModel.userMessage.value = msg
 		alert match {
-			case 0 => CiteMainModel.userAlert := "default"
-			case 1 => CiteMainModel.userAlert := "wait"
-			case 2 => CiteMainModel.userAlert := "warn"
+			case 0 => CiteMainModel.userAlert.value = "default"
+			case 1 => CiteMainModel.userAlert.value = "wait"
+			case 2 => CiteMainModel.userAlert.value = "warn"
 		}
 		js.timers.clearTimeout(CiteMainModel.msgTimer)
-		CiteMainModel.msgTimer = js.timers.setTimeout(20000){ CiteMainModel.userMessageVisibility := "app_hidden" }
+		CiteMainModel.msgTimer = js.timers.setTimeout(10000){ CiteMainModel.userMessageVisibility.value = "app_hidden" }
 	}
 
 
 	def loadLocalLibrary(e: Event):Unit = {
 		val reader = new org.scalajs.dom.raw.FileReader()
-		val delimiter:String = {
-			val delimiterChoice = js.Dynamic.global.document.getElementById("app_filePicker_delimiter").value.toString
-			if (delimiterChoice == "TAB" ){ val d = "\t"; d } else { val d = "#"; d }
-		}
 		CiteMainController.updateUserMessage("Loading local library.",0)
 		reader.readAsText(e.target.asInstanceOf[org.scalajs.dom.raw.HTMLInputElement].files(0))
 		reader.onload = (e: Event) => {
 			val contents = reader.result.asInstanceOf[String]
-			CiteMainController.updateRepository(contents,delimiter)
+			CiteMainController.updateRepository(contents)
 		}
 	}
 
@@ -80,21 +87,45 @@ object CiteMainController {
 			js.Dynamic.global.document.getElementById("tab-1").checked = true
 	}
 
+	def hideTabs:Unit = {
+
+	  CiteMainModel.showTexts.value = true
+	  CiteMainModel.showNg.value = true
+	}
+
+	def checkDefaultTab:Unit = {
+		if (CiteMainModel.showTexts.value) {
+			js.Dynamic.global.document.getElementById("tab-1").checked = true
+		} 
+	}
+
+	def clearRepositories:Unit = {
+		O2Model.textRepository = null
+	}
+
+
+	// Reads CEX file, creates repositories for Texts, Objects, and Images
+	// *** Apropos Microservice ***
 	@dom
-	def updateRepository(cexString: String, columnDelimiter: String = "\t") = {
+	def updateRepository(cexString: String) = {
+
+		hideTabs
+		clearRepositories
 
 		try {
-			val repo:CiteLibrary = CiteLibrary(cexString, columnDelimiter)
-			println("got here okay.")
+
+			val repo:CiteLibrary = CiteLibrary(cexString, CiteMainModel.cexMainDelimiter, CiteMainModel.cexSecondaryDelimiter)
 			val mdString = s"Repository: ${repo.name}. Library URN: ${repo.urn}. License: ${repo.license}"
+			var loadMessage:String = ""
 
 			repo.textRepository match {
 				case Some(tr) => {
-					CiteMainModel.currentLibraryMetadataString := mdString
-					CiteMainController.updateUserMessage(s"Created new corpus. ${mdString}",0)
+					CiteMainModel.showTexts.value = true
+					CiteMainModel.showNg.value = true
+					CiteMainModel.currentLibraryMetadataString.value = mdString
 					O2Model.textRepository = tr
-					CiteMainController.updateUserMessage(s"Updated text repository: ${ O2Model.textRepository.catalog.size } works.",0)
-
+					CiteMainController.updateUserMessage(s"Updated text repository: ${ O2Model.textRepository.catalog.size } works. ",0)
+					loadMessage += s"Updated text repository: ${ O2Model.textRepository.catalog.size } works. "
 					O2Model.updateCitedWorks
 					NGModel.updateCitedWorks
 					NGController.clearResults
@@ -105,13 +136,17 @@ object CiteMainController {
 					NGController.preloadUrn
 				}
 				case None => {
-					CiteMainController.updateUserMessage("Chosen repository does not seem to include a TextRepository",2)
+					loadMessage += "No texts. "
 				}
 			}
 
+			checkDefaultTab
+
+			CiteMainController.updateUserMessage(loadMessage,0)
+
 		} catch  {
 			case e: Exception => {
-				CiteMainController.updateUserMessage(s"""${e}. You might check to be sure you specified the correct delimiter (<tab> or "#").""",2)
+				CiteMainController.updateUserMessage(s"""${e}. Invalid CEX file.""",2)
 			}
 		}