+++ src/main/scala/citeappscalajs/NGModel.scala
@@ -9,11 +9,12 @@ import org.scalajs.dom.ext._
 import org.scalajs.dom.raw._
 import edu.holycross.shot.cite._
 import edu.holycross.shot.ohco2._
-import edu.holycross.shot.citeenv._
+import edu.holycross.shot.citeobj._
 
 import scala.scalajs.js.annotation.JSExport
+import js.annotation._
 
-@JSExport
+@JSExportTopLevel("citeapp.NGModel")
 object NGModel {
 
 	var msgTimer:scala.scalajs.js.timers.SetTimeoutHandle = null
@@ -87,26 +88,26 @@ object NGModel {
 	@dom
 	def updateShortWorkLabel = {
 		if ( O2Model.textRepository == null){
-					NGModel.shortWorkLabel := "- no selected text -"
+					NGModel.shortWorkLabel.value = "- no selected text -"
 		} else {
-			val longS:String = O2Model.textRepository.catalog.label(NGModel.urn.get)
+			val longS:String = O2Model.textRepository.catalog.label(NGModel.urn.value)
 			if (longS.size > 50){
 				val shortS:String = longS.take(24) + " … " + longS.takeRight(23)
-				NGModel.shortWorkLabel := shortS
+				NGModel.shortWorkLabel.value = shortS
 			} else {
-				NGModel.shortWorkLabel := longS
+				NGModel.shortWorkLabel.value = longS
 			}
 		}
 	}
 
 	@dom
 	def updateCitedWorks = {
-		NGModel.citedWorks.get.clear
+		NGModel.citedWorks.value.clear
 		NGController.clearResults
 		NGController.clearInputs
 		// N.b. The textRepository remains with the Ohco2 Model.
 		for ( cw <- O2Model.textRepository.corpus.citedWorks){
-			NGModel.citedWorks.get += cw
+			NGModel.citedWorks.value += cw
 		}
 	}
 
@@ -152,7 +153,7 @@ object NGModel {
 
 /* String and Token Finding */
 def findString(urn:CtsUrn, s:String):Corpus = {
-		val tempCorpus = O2Model.textRepository.corpus ~~ NGModel.urn.get
+		val tempCorpus = O2Model.textRepository.corpus ~~ NGModel.urn.value
 		val foundCorpus = tempCorpus.find(s)
 		foundCorpus
 }